pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = ""
authors = ["13dagger <13dagger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
